v1.8 (Internet sample updates)
I found that both the Simple and Advanced Internet Shield samples were
crashing.   It ended up being that I had to Initialize the HttpRequest
variables after the OneSheeld.begin() call.  Which required initializing
a pointer, which in-turn required some other changes.  But all changes
were because of the initialization differences.   If I were to guess,
I'd say the crash is because the HttpRequest initialization was using
something that required something else that hadn't been initialized
because it got somewhere before the .begin call made it to the point to
initialize what was needed.  But I really haven't investigated far
enough to know for sure (but I think it's because the pcDuinos are a lot
faster of a computer than the Arduinos are) ... but I could be way off
too!   What I do know, is that with these changes, both of the Internet
Shield samples no work!!!!   Note: This is required for the IDE version
of the samples too!!!!